// star-wars-app/models/starship.dto.ts
import Basic from "./basic.dto";

export default interface Starship extends Basic {
  MGLT: string;
  cargo_capacity: string;
  consumables: string;
  cost_in_credits: string;
  created: string;
  crew: string;
  edited: string;
  hyperdrive_rating: string;
  length: string;
  manufacturer: string;
  max_atmosphering_speed: string;
  model: string;
  name: string;
  passengers: string;
  films: string[];
  pilots: [];
  starship_class: string;
  url: string;
}